requirements.txt
paho-mqtt
pylast
